requirements.txt
streamlit
tensorflow
tensorflow-datasets
Pillow
matplotlib
pickle